Fixing IoChannel.messageType usage.

+++ src/kernel/NanoOsTypes.h
@@ -168,7 +168,7 @@ typedef struct NanoOsFile {
 /// @param messageType The type of message to send to the process.
 typedef struct IoChannel {
   ProcessId pid;
-  uint8_t   messageType;
+  int64_t   messageType;
 } IoChannel;
 
 /// @struct FileDescriptor
